contact application finished

+++ lib/ui/single_contact/single_contact_screen.dart
@@ -6,12 +6,12 @@ import 'package:zoom_tap_animation/zoom_tap_animation.dart';
 import '../../data/db/local_database.dart';
 import '../../model/contact_model.dart';
 import '../contact_screen/contact_screen.dart';
-
+import 'package:url_launcher/url_launcher.dart';
 class SingleContactScreen extends StatefulWidget {
-  const SingleContactScreen({Key? key, required this.id}) : super(key: key);
+  const SingleContactScreen({Key? key, required this.id, required this.listening}) : super(key: key);
 
   final int id;
-
+  final VoidCallback listening;
   @override
   State<SingleContactScreen> createState() => _SingleContactScreenState();
 }
@@ -22,6 +22,7 @@ class _SingleContactScreenState extends State<SingleContactScreen> {
   String name = "";
   String surname = "";
   String phone = "";
+  String photo = "";
   late ContactModelSql contactModelSql ;
 
 
@@ -31,14 +32,13 @@ class _SingleContactScreenState extends State<SingleContactScreen> {
     name = contactModelSql.name;
     surname = contactModelSql.surname;
     phone = contactModelSql.phone;
+    photo = contactModelSql.photo;
     setState(() {});
   }
 
   @override
   void initState() {
     _updateContacts();
-
-
     super.initState();
   }
 
@@ -47,16 +47,11 @@ class _SingleContactScreenState extends State<SingleContactScreen> {
     return Scaffold(
       appBar: AppBar(
         leading: IconButton(onPressed: (){
-
-          Navigator.pushReplacement(context, MaterialPageRoute(builder: (context)=>const ContactScreen()));
+          widget.listening.call();
+          Navigator.push(context, MaterialPageRoute(builder: (context)=>const ContactScreen()));
         }, icon: const Icon(Icons.arrow_back,color: Colors.black,)),
         backgroundColor: Colors.white,
         title:const Text("Contacts",style: TextStyle(fontSize: 20,fontWeight: FontWeight.w500,color: Colors.black),),
-        actions: [
-          IconButton(onPressed: (){}, icon:const Icon(Icons.search,color: Colors.black,),),
-          IconButton(onPressed: (){}, icon:const Icon(Icons.more_vert,color: Colors.black,),),
-
-        ],
       ),
       body: Padding(
         padding: const EdgeInsets.symmetric(horizontal: 15,vertical: 30),
@@ -66,7 +61,14 @@ class _SingleContactScreenState extends State<SingleContactScreen> {
             Row(
               children: [
                 const SizedBox(width: 130,),
-                SvgPicture.asset("assets/svg/big_account.svg"),
+                Container(
+                  padding: EdgeInsets.all(10),
+                  margin: EdgeInsets.all(10),
+                  decoration: BoxDecoration(
+                    shape: BoxShape.circle,
+                  ),
+                  child: Center(child: Image.asset(photo,width: 80,height: 80,),),
+                ),
                 const Spacer(),
                Column(
                  children: [
@@ -74,10 +76,34 @@ class _SingleContactScreenState extends State<SingleContactScreen> {
                   Row(
                     children: [
                       ZoomTapAnimation(
-                        onTap: (){
-                          LocalDatabase.deleteContact(contactModelSql.id!);
-                          Navigator.pushReplacement(context, MaterialPageRoute(builder: (context)=>const ContactScreen()));
-                        },
+                          onTap: () {
+                            Future.delayed(
+                                const Duration(seconds: 0),
+                                    () => showDialog(
+                                  context: context,
+                                  builder: (context) => AlertDialog(
+                                    title: const Text('Delete everything?'),
+                                    content: const Text(
+                                        'Are you sure you want to remove everything'),
+                                    actions: [
+                                      TextButton(
+                                          onPressed: () {
+                                            Navigator.pop(context);
+                                          },
+                                          child: const Text('No')),
+                                      TextButton(
+                                          onPressed: () {
+                                            LocalDatabase.deleteContact(contactModelSql.id!);
+                                            Navigator.pushReplacement(context, MaterialPageRoute(builder: (context)=>const ContactScreen()));
+                                          },
+                                          child: const Text(
+                                            'Yes',
+                                            style: TextStyle(color: Colors.red),
+                                          )),
+                                    ],
+                                  ),
+                                ));
+                          },
                           child: SvgPicture.asset("assets/svg/delete.svg")),
                       const SizedBox(width: 15,),
                       ZoomTapAnimation(
@@ -103,11 +129,22 @@ class _SingleContactScreenState extends State<SingleContactScreen> {
             const SizedBox(height: 40,),
             Row(
               children: [
-                Text("+998${phone[0]}${phone[1]} ${phone[2]}${phone[3]}${phone[4]} ${phone[5]}${phone[6]} ${phone[7]}${phone[8]}",style: const TextStyle(fontSize: 16,fontWeight: FontWeight.w500,color: Colors.black),),
+                Text("+998 ${phone[0]}${phone[1]} ${phone[2]}${phone[3]}${phone[4]} ${phone[5]}${phone[6]} ${phone[7]}${phone[8]}",style: const TextStyle(fontSize: 16,fontWeight: FontWeight.w500,color: Colors.black),),
                 const Spacer(),
-                SvgPicture.asset("assets/svg/phone.svg"),
+                ZoomTapAnimation(
+                    onTap: ()async{
+                      await launchUrl(Uri.parse("tel:+998$phone"));
+                    },
+                    child: SvgPicture.asset("assets/svg/phone.svg")),
                 const SizedBox(width: 15,),
-                SvgPicture.asset("assets/svg/sms.svg"),
+                ZoomTapAnimation(
+                    onTap: ()async{
+                      String uri = "";
+                      uri ="sms:+998$phone?body=";
+                      // ignore: deprecated_member_use
+                      await launch(uri);
+                    },
+                    child: SvgPicture.asset("assets/svg/sms.svg")),
               ],
             )
           ],
